Add .idea to .gitignore and set up resource pack deployment
The .idea folder, generated by the IntelliJ IDEA IDE, has been added to .gitignore to prevent it from being tracked. In addition, a new GitHub action workflow has been created for resource pack deployment. This workflow is triggered on push or workflow dispatch events, and it downloads the resource pack from the specified GitHub repository, then uploads it to the specified SFTP server.